Move Embedded operate to persistence module.
